refactor: move api test code
